+++ tap_mysql/sync_strategies/common.py
@@ -81,7 +81,11 @@ def generate_select_sql(catalog_entry, columns):
         # if the column format is binary, fetch the values after removing any trailing
         # null bytes 0x00 and hexifying the column.
         if 'binary' == property_format:
-            escaped_columns.append(f'hex(trim(trailing CHAR(0x00) from {escaped_col})) as {escaped_col}')
+            escaped_columns.append(
+                f'hex(trim(trailing CHAR(0x00) from {escaped_col})) as {escaped_col}')
+        elif 'spatial' == property_format:
+            escaped_columns.append(
+                f'ST_AsGeoJSON({escaped_col}) as {escaped_col}')
         else:
             escaped_columns.append(escaped_col)
 
@@ -115,7 +119,7 @@ def row_to_singer_record(catalog_entry, version, row, columns, time_extracted):
         elif 'boolean' in property_type or property_type == 'boolean':
             if elem is None:
                 boolean_representation = None
-            elif elem == 0:
+            elif elem == 0 or elem == b'\x00':
                 boolean_representation = False
             else:
                 boolean_representation = True